requirements.txt
aiogram==3.13.0
httpx==0.27.2
yt-dlp>=2024.10.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
